Allow users to create taxonomy projects
This adds interface for users to create new projects. This should make
it easier to test by content designers, because there's no developer
involvement.

Because the todos are created in the request, this might not work for
large spreadsheets (the page might time out). In that case we can still
use the rake task.