Change the layout for the audit part of the CPM
Add a layout for the audit related pages, with a different app title
and page title, as well as a different way of laying out the content.

Move the tabs on the index and report page to encompass both the
filters and the main page content, as there is a plan to have the
filters change when you change tab.

Also remove the page headings, now that the app title and page title
are set.