Provide a checkbox to toggle primary/all orgs. filter
https://trello.com/c/MzAVD2dG/320-2-update-org-filter-to-toggle-primary-filter-on-off

The hidden field means the URL param is set to
false when not explicitly overriden by the
checkbox. We need to do this because we want to
checkbox to default to checked.

The count of content for each organisation will
depend on whether the checkbox is selected during
page load. If the user changes the checkbox, those
numbers won’t be updated immediately.

We’d need to progressively enhance the form with
JS to solve this problem.