Add a secondary sort by the id in the Search
This will ensure that the search results come back in a deterministic
order.

Consistency is important for users of the application, but also has
technical implications on pagination in the query. When using LIMIT
and OFFSET, if there are say two entries that sit on the boundary
between pages, with identical positions with respect to the primary
sort (e.g. they have the same number of page views), then the database
can return one of the items for both pages, and the other item on
neither page.

The above scenario is also how this issue was discovered.