Fix moving across page boundaries with the Next link
Expand the scope for finding the next item by one, and update the page
parameter for the filter parameters that are used in the next link.

Previously the Next link wouldn't appear on a page for a audit, if
that audit sat at the end of a page (a subset of all the audits,
specified by an offset and size).

Previously, the responsibility for identifying the next page sat
within the ContentItem model. It however, only looked at content items
on the current page, so it couldn't find the item on the next page.

Finding the item on the next page can be achived by extending the size
of the page by one, only when looking for the next item. This however
isn't sufficient, as the page parameter in the request must be set
correctly, otherwise the problem is just moved to the next page. When
moving across page boundaries, the page parameter must also be
incremented, such that the scope of content items includes the current
item.

As the ContentItem model isn't involved in pagination, this is better
handled in the controller.