Separate Search and Query concepts
The concepts of a “search” and a “query” were very tightly coupled, to
the point where a Search is initialized, and the query mutated through
the Search object. The Search is then executed, and the query could be
mutated even after the execution, leading to the possibility of
having a query that does not match the results.

Further, the search object was abused to create and mutate clones of its
own query to perform separate searches.

This change makes Query objects immutable, and moves their creation
logic into a separate Builder. A Search is now initialised with a query,
which cannot be changed, so a search will always represent a single
search.