Fetch next item through search query
The previous approach of fetching the “next” content item relied on
navigating an underlying list of results. This can be problematic,
because if the list of results is paginated, you can’t fetch the next
item from the next page of results.

Further, the results may also grow stale. For example, if searching for
non-audited content items, and someone audits a content item, then the
results change underneath us.

This change instead requests the next result directly through a SQL
query, rather than relying on traversing a list of results, which is
guaranteed (apart from certain race conditions) to return the next item
that needs looking at.