Change slugs to use content_id instead of id
We refer to content items by their content_item_id in most GOV.UK apps
so we should be consistent. Here we’ve switched the URL param when
referring to content items.

We add a method to ContentItem to support slugs that use the content_id
instead of internal id more naturally.