Stabilise sort order
Our sort was previously unstable. That is, if we sorted on - for
example - Page Views, if two results had equal page views, they may
appear in any random order.

This change breaks all ties on the content items’ base paths, which are
unique, and hence guarantees that a given sort order will always return
items in the same order.