Add Google Tag Manager script
This change adds a script to enable tracking through Google Tag Manager.

Notably, we have not included the `noscript` `iframe` mentioned in the
setup documentation; this `iframe` is for tracking users with
JavaScript disabled. However, it requires extra setup, and we currently
do not anticipate needing to track users without JavaScript.