Remove filter counts
The drop-down filters and the inventory manager had counts of content
next to them. For example, in the “Organisations” filter, you might
see “Department for Education (12)”, which would mean that DfE has
12 content items that match your other (non-organisational) filters.

While this is a nice-to-have, there is no confirmed user need for it,
and it complicates the code, because this behaviour is tied heavily
to the Search code (which we hope to refactor).