Add Content Query and Audit Filtered Query and use in the Audit controller
This change introduces a `Content::Query` class, which is intended to
be part of the replacement for the entire `Search` class (and its
subclasses).

This change begins the process of isolating the concept of content-only
querying from the concept of filtering the content based on joins with
other tables (ie Audits).