Fetch content after another content item
This change adds the ability to fetch the content item after a given
content item, according to the sort criteria provided in a query.

This is particularly useful when trying to pick the “next” item when
viewing a single content item.

The previous approach to this used the list of content items on a page,
but that approach breaks when viewing the last item on a page.