Fetch Content Items with Locale
Some content items (approx. 1,600) do not have an “en” locale, which is
the assumed default when querying the Publishing API.

Asking for one of these content items by content ID, but without a valid
locale will result (correctly) in a 404 error from the Publishing API.

In order to address this, when fetching all content IDs from the
`/v2/editions` endpoint, we will also fetch the corresponding locales.
If the “en” locale is present, we use that by default. If it is not
present, we take the first arbitrary locale, so that the content item
is definitely included in the database.

We currently only accept one locale to avoid the race condition of
updating two content items with the same content ID at the same time.