Add override to running code coverage
We don't always want to generate the coverage report when we are
running the tests locally during development, as it slows down the tests.

The coverage report will not be generated when running guard. Running
default rake or just rspec will still generate the report, but these can
be overridden with an additional env var.

```
NO_COVERAGE=true bundle exec rspec
```

We can even alias this.