Return 404 for non-existent content items
The `find_by` method returns `nil` if a model cannot be found. We then
try to run `decorate` on `nil`, which causes a `500` error. This is
incorrect, and we should return a `404`.

Using `find_by!` will instead raise an `ActiveRecord::RecordNotFound`
error, which returns a `404` status, and a default “Not Found” page.