Pick a random next item for term generation
When 2 or more content designers are working on the same project, they
will constantly be working on the same content item. We'd like to solve
this with an allocation system (where we assign users to a set of
things to review), but that isn't ready yet. By picking a random todo
item we avoid situations where users see the same page and overwrite
each others changes.

https://trello.com/c/92bRRnvm