Remove dependency between tests and a Global Seeder
Tests should not rely on global fixtures set up in a Seeder. The more
explicit the fixtures for a test the better. In this case, it contained
an explicit role that was needed for the test 
`spec/features/inventory/inventory_spec.rb`

I have removed the seeder code to setup the user, and I have setup 
a `config.before(type: feature)` for RSpec.feature tests.