Do not order Content Items in CSV reports
Ordering Content Items approximately doubles the length of time required
to execute the query.

When selecting a large number of columns, as we do for generating a
report, this can slow down the query enough to cause 504 Gateway
Timeout errors.

This change removes the ordering from CSV report queries in order to
speed up its performance and prevent 504 errors.