Fix applying csv filtering test
The test for checking whether filtering is applied to the csv export was
looking for a non-existent example. This change makes the test explicitly
check for what should and shouldn't be in the export when filtering is
applied.

Prior to this the test was only checking for non-existant content and
would always have passed.