Amend guidance docs to not refer to spreadsheet
This guidance doc is linked to from within the audit tool and is in the
context of the audit tool, but was originally written for the
spreadsheets and had not been updated.

Ahead of the user research test sessions we have this week, I have made
some minor edits. We will revisit this content with a content designer
in the near future.