Fix Lint warning: Metrics/BlockLength
A recent upgrade of Rubocop brought a new rule into play, that blocks
should be no longer than 25 lines.

Now, RuboCop is pretty smart, but it doesn't know what is a DSL and not,
so we can't automatically ignore them. One could argue that we could
exclude DSL entry methods of popular frameworks, like Rails routes and
RSpec specs. The reasons for not doing so are primarily:

False negatives. Any class can implement a method, taking a block, with
the same name. RuboCop is a Ruby analysis tool, and shouldn't really
know about external libraries. (Excluding the /spec directory is a
courtesy until we have a proper extension system, and this can be
handled by the rubocop-rspec gem.)