Show only audible formats
Not all formats are auditable by content auditors, so we need to exclude
them. We have decided to define a `Plan` and audit that determines what
subset of content items are auditable.

As of now, we are only excluding by format (document type), but in the
long run we plan to extend the plan definition with subscriptions and
individual content items.