Do not override GA attributes when importing
We are currently overwriting the GA attributes that are stored in the
Content::Item. We plan to move those attributes out of Content::Item
soon, but in the meantime we should skip them from the importing process.

This issue has raised because the GA data in integration is different
thank in production.