Filter content allocated to other content auditors
We have decided to allow auditors to filter content that is allocated
to other content auditors. The approach that we have decided to follow
is to get the list from the table `Users` that is local to the CPM, and
is maintained automatically by the sign-on gem.

We have explored other options, like retrieving the list from Sign-on,
but it seems to be too much work for not much benefit.