Filter Content Auditors when allocating content
Auditing is performed in teams with content auditors that belong to the
same organisation. It is preferred then, to show only the auditors of
the team when it comes to allocating and unallocating content.

It is also desired not to list all the users that belong to an
organisation, but to be more restrictive, and list only the users that
has content allocated to them.