Track `select` element change events
By default, Google Tag Manager does not capture `change` events. We
would like to track the values selected in `select` elements when they
are changed.

This change adds a single JavaScript listener that will forward `change`
events to GTM if the element that was changed was a `select` element.
This event will include the value of the element.