Add Tracking ID attributes to select elements
For analytics, we need a stable element identifier that will not change,
even if other parts of an element might change (such as ID, or classes).

This change adds a tracking identifier to each `select` element, which
can be used in Google Tag Manager to filter events for specific
`select` fields.