Change Audit Status filter to radio buttons
Radio buttons are more accessible, and generally more apparent than
dropdown menus. This change replaces the old `select` element with a set
of radio buttons for selecting the audit status.

This change also reorders the audit status options, and uses “Not
audited” as the default, both in the view, and in the controller.