Present all organisation users in dropdowns
When picking colleagues from dropdowns, we currently only show users
with content allocated to them. This can make for a confusing user
experience when trying to - for example - allocate content to users
who have never had content allocated to them, or when trying to view
content allocated to that user.

This change removes that restriction. The users in our database are only
the users who have logged into the application, which will also be
restricted by Signon permissions. This will naturally limit the number
of other users presented in the dropdowns, so it is assumed that we
do not need to restrict users in this way.