Synchronizes UI components in allocation
When the user updates the number of elements to allocate/unallocate via
the input box, we need to select the checkboxes manually. So they are
synchronised.

When the user clicks on the checkboxes we need to count the number of
checkboxes that are selected and update the input for the batch size.