Make filter available to Audit views and helpers
The query string is not necessarily indicative of the audit view state.
For example, when landing on `/audits`, we have no query string, but
we have some “non-blank” values in our filter (ie we are filtering to
unaudited content assigned to “Me”).

In order to make decisions in the view about the current state of the
filter, we should make this filter available to the view, and to helper
methods, by making the filter itself available as a memoized helper
method.