Let users clear the first organisation filter with an “X” button
We currently hide the “X” remove organisation button if there is only
one organisation filter. However, users cannot clear this selection, so
we should add this functionality back.

This change unhides the “X” button if there is only one organisation,
and fixes a bug that means clearing the input opens the options
without being able to dismiss them.