Allow different tabs to have different default filters
As we continue to diverge the sidebars used in the “Audit content”
and “Assign content” tabs, we have a need to set different default
values for the filters in these views. These defaults reflect the
actions users will need to carry out on these pages.

In order to allow different default values, this change does the
following:

- Remove all knowledge from the `Filter` class about default values,
  and remove any special meaning assigned to “blank” values
- Move default filter settings into the controllers
- Fix the filters to rely on the `filter` object to set their
  `selected` values, because `params` does not necessarily reflect the
  state of the filter

This change also fixes the previous bug whereby the “Assigned to” filter
would show “Me” when landing on the “Audit content” page, but it would
show content assigned to anyone. Since this bugfix changes the content
shown by default on the “Audit content” page, a number of assumptions
have had to be adjusted in corresponding tests.