Remove automatic user setup from feature testing
We currently "hide" the setup of our logged-in user in a `before`
block. This can obscure logic based on the current user, and hides
the assumptions made about them.

It also makes it harder to set the current user up with - say - a
particular organisation, which may be needed for testing.

This change removes the automatic setup, forcing each feature test to
declare its own logged-in user, along with any assumptions and extra
setup required (eg a particular name, which might be checked in banner
messages, or a particular organisation, which may be used for content
or selected from dropdowns, etc.).