Move defaults out of Filter class
The `Filter` class currently has an understanding of default values,
and assigns special meaning to blank values for some fields.

We want to move towards different views having different default values
for particular filter fields, so in preparation for this, we need to
remove any knowledge of defaults of special blank values from the
Filter, and move this behaviour up into the controllers.