Render allocation hidden fields directly from the filter
We currently render a set of hidden `input` fields when allocating
content. These fields are used to construct the redirect URL after
content allocation. However, they are derived from the query string.

The query string does not reliably indicate the state of our filter
(ie when landing on the page cold), and we should query the filter
directly.

This change constructs these fields directly from the `filter` object.
In order to be able to do this, the `Filter` object has also been
turned into a `Struct` for its built-in `.to_h` method.

As a by-product of this refactor, the logic for constructing and
splitting the sort strings has been moved out into its own class, with
improved test coverage.