Pull sidebar filter values directly from the filter object
The sidebar filter values are currently populated from the query
string parameters. However, these parameters are not necessarily a
valid indicator of the state of the filter (ie when landing cold on the
page with no query string).

This change instead gets the current filter status directly from the
`filter` object, which is the object used to render the view, hence
keeping the sidebar in sync with the view.