Stop query string propagating between tabs
In order to help users focus on the task in each tab, we have now
changed each one to have sensible default filter values.

If we propagate the user's filter values across tabs, then these
defaults will not be seen if the user has already interacted with the
filters.

This change removes the query string propagation, so that users will be
presented with sensible defaults every time they change tab.