Log certain Sidekiq errors only on their final retry
Some of our Sidekiq errors arise from interactions with 3rd party APIs,
and failures beyond our control are to be expected from time to time.
We do not want to be notified every time we get one of these errors;
we just want to be notified the final time the job fails.

This change updates our `ApplicationJob` to perform its tasks inside a
block that will be rescued when certain errors raise. It will then
re-raise a `RetryableError`, which will be ignored by Sentry.

When the retries on a job are exhausted, the error will finally be sent
to Sentry.

Raven is Sentry's API offering. Documentation on the Ruby implementation
can be found [here][raven].

We thought about solving this using Raven's `should_capture` method,
which would enable us to filter errors base on arbitrary conditions,
but that would have require some error parsing. The approach taken here
is more explicit about the expected errors, and requires fewer moving
parts.

[raven]: https://docs.sentry.io/clients/ruby/