Assign jobs to their own queues with different concurrencies
We have different concurrency requirements for our different Sidekiq
jobs:

- The Content Item import job takes hours to run, so it needs concurrent
  runners, but not too many so that we avoid hammering the Publishing
  API.
- We keep getting [concurrency errors][errors] from Google when we query
  for analytics data, so we need to remove all concurrency from those
  jobs.

In order to have different concurrency settings, we will spin up two
separate Sidekiq processes that will each handle their own queue.

This change adds the configuration for these processes, and configures
the jobs to assign themselves to the correct queue.

Note that in a future commit, we should remove the old `sidekiq.yml`
configuration. It is left in in this commit to avoid errors when
deploying.

Also note that neither of these jobs will run until the corresponding
Puppet change is merged.

[errors]: https://sentry.io/govuk/app-content-performance-manager/issues/352869156/