Handle `nil` rows in Google Analytics responses
When we fetch pageview data from Google Analytics, it returns a response
object. If none of the queried base paths have any pageviews, then
the response object's rows will be `nil`, rather than an empty array.

This case may arise if we query a batch of uncommonly-accessed documents,
such as a [batch of announcments][sentry].

In order to guard against `nil` errors, this change checks that the
`rows` field exists, and - if it isn't - will provide an empty array.

[sentry]: https://sentry.io/govuk/app-content-performance-manager/issues/353541553/events/7483935478/