Remove GTM assertons from analytics specs
This removes the assertion for a Google Tag Manager property from our
test. We have seen this cause failures under CI which I suspect is
caused by a change in the behavior of GTM in that environment.

This change isolates us from behavior introduced by GTM. This behavior
is also not one we own so not one we should test here.