Allow blank organisations even with defaults
This change addresses a bug we have:

1. Visit the "Allocation" page
2. The organisation filters should default to your own organisation
3. Clear the organisation filter
4. Apply filters
5. The page is still filtered by your own organisation

This is because we have no way of determining the difference between
a user who has just landed on the page "cold" (ie with no filters set),
or a user who has actively deselected all the organisations (or selected
the blank organisation).

This change addresses this issue by adding a hidden `<input>` field
just before the `<select>` element. This has the effect of including a
blank organisation array whenever the filters are applied from the page,
even if no options are selected.

We can then differentiate in the backend on whether the request was
actively for a blank organisation filter (the `params` key will be
present), or just a "cold" visit (the `params` key will not be present)
not be present.

This solution has already been employed for our
[primary org filter][primary org filter], and is also the method
employed by [Rails itself][rails].

[primary org filter]: https://github.com/alphagov/content-performance-manager/blob/a46bc1ead0b0f6858f2944d14c8d7cdae6abd8d8/app/views/audits/common/_primary.html.erb#L2
[rails]: http://api.rubyonrails.org/v5.1/classes/ActionView/Helpers/FormOptionsHelper.html#method-i-select