Ensure filter defaults are applied on first interaction
When you first visit one of the audit tabs, the querystring does not
contain any filter state. This means default filters are only applied
if you interact with an element on the page that has URLs appended
with the filter defaults.

This fixes that so that the defaults are applied correctly.