Tie report headers to their values
At the moment, our report headers are not coupled to the values returned by a `ReportRow`. This can be dangerous when adding, removing, or reordering data fields, because the headers may not match the content. This change moves the headers right next to the values they represent, in order to promote consistency between the two.