Show 100 items per page
This change shows 100 items per page on both the audit and allocation pages. This is to address a user need to be able to quickly find the content they are looking for. Showing more items should reduce the amount of paging they have to do, and make it easier to find their content.