Look up Users' organisation titles from Content Items
We currently try to infer the title of a user's organisation by [`titleize`ing][1] their organisation slug. However, this doesn't work in all cases. For example, HMRC has the slug `/hm-revenue-customs`, which titleizes to "Hm Revenue Customs", when the actual title of the organisation is "HM Revenue & Customs" (note the incorrect capitalization of "HM" and the missing ampersand).

This change removes the titleization, and instead explicitly looks up the organisation title from the organisation's Content Item. This is done for both auditors and assignees.

[1]: http://api.rubyonrails.org/v5.1/classes/ActiveSupport/Inflector.html#method-i-titleize