Add rake task to deploy current branch to Heroku
Add a `Rake` task to deploy branches to Heroku so we can fact-check
features and user-research them independently of the status of our
master branch:

1. The master branch is always in a deployable state, so we decrease the
 risk of being blocked with non-approved features.
2. We don’t have to write production-ready code to user research new
features.

Regarding the Rake task:

1. The application name depends on the PR number, so this is easier to
locate in Github.
2. It uses different `git` remotes (not always `heroku`) as we deploy
multiple branches and each one of them will spin up a new `Heroku`
application.
3. We rely on our codebase to retrieve the list of content items. The
Queries are translated into SQL by using `to_sql` on `ActiveRecord`
objects.
4. We are not able to use `pg_dump` because this command performs a full
export of the database, and we need to filter the number of content
items that we want to export.
5. We `force-push` in each deployment as we want to reuse the same code
 for redeployments.