Prevent failure if a session is attached to a removed user
If a user is removed from the database and the `id` is referenced by
a user session, then the prototype start raising 500s errors.

Clearing the session does the trick.

This is an edge case, but it seems sensible to handle it here to ensure
that user research is not interrupted for technical reasons.