Load Report Rows on the fly
The CSV export was broken due to an unexpected Array object instead of
a Hash, returned by the Audits::Report `rows` method.

A new method of retrieving reports rows was introduced, but this is not
supported by the old values anymore.

This problem highlighted that the overall approach used for report
generation is difficult to maintain.

In order to avoid these problems we are going to load the reports' rows
on the fly and to not use a dedicated DB table.