Fix batch assignment from page 2 onwards
This fixes the logic we have to calculate which content items to
allocate when a batch size is specified that is greater than the 100
items we display on a page and when the assignment is instigated from
page 2 or greater.

We calculate an offset value which is the total number of content
items that would have been displayed on all previous pages. We then
assign X number of content items from there onwards.