Refactor "filter content audited for anyone" test
This refactors the second test in the Filter feature test.

Because the test file is so large, we will refactor one test at a time
in separate commits to help with readability of changes.

This changes the style of how the test is written to add clear
separation between the Arrange, Act, Assert parts of the test.

See https://about.futurelearn.com/blog/how-we-write-readable-feature-tests-with-rspec