Refactor tests to not rely on HTML classes
As suggested by @KyleMacPherson, we are using the `data-test` attribute
to assign to our markup so that our feature tests can select elements.
By not relying on classes, we are able to separate the concern of
classes so that they can remain specific to styling. This means that if
the styling and classes needed to change, the tests would not break.