Ignore drafts when fetching content items
We currently have a mismatch between the Audit Tool and the live GOV.UK
site when a draft of a content item exists.

If a draft exists, the Audit Tool will import the draft, and display
that, rather than the latest published edition (which is shown on
GOV.UK).

This change adds a method to our Publishing API client to fetch the
latest published edition, or to return `nil`, so that the content we
display in the Audit Tool will match that on GOV.UK.