Single-column audit page layout
We wish to add a preview of the content item being audited to the side
of the audit page. In order to prepare for this, this change collapses
the audit page into a single-column layout.

To achieve this single-column layout, the metadata is to be moved from
its own column to a grid underneath the description.

The grid layout is achieved by separating the concern of the grid
content from the mechanic of creating the Boostrap grid.