Add a content preview to the audit page
This change adds a preview of the content being audited to the right of
the audit questions. It makes use of the iframe proxy that is already
in the application to get around the fact that GOV.UK does not allow
its content to be displayed in an iframe.

The iframe is sized using flexbox, which enables us to fill the pane
with the iframe without using JavaScript. However, older versions
of Internet Explorer are not compatible with flexbox. In these cases,
the iframe falls back to a fixed height.

A potential alternative to this flexbox approach could have been to use
`position: fixed` and fix the two panes to the two halves of the screen,
but then we would lose the nice responsive behaviour that Bootstrap
gives us, enabling us to collapse the iframe on mobile devices.