Open iframe links in a new tab
This change augments the behaviour of the iframe proxy. The iframe
proxy previously rewrote all GOV.UK links to pass through the proxy
instead. This would enable the user to click links in the iframe, and
have the next page load in the iframe via the proxy.

Rather than keep the user in the iframe, we now open links in a new tab,
which makes getting back to the previous page more intuitive (rather
than figuring out how to navigate "Back" in an iframe).

Because we now open all links in a new tab, the only links we have to
rewrite are relative paths, which are prepended with `https://www.gov.uk`.

The final change made is to concatenate the body fragments and parse
the entire HTML document as HTML. Regexing HTML is unreliable and hard
to read, and is actually broken when trying to use it on body fragments,
since HTML tags may span multiple fragments.