Fix content preview after saving an audit
After saving an audit, if there is no content item to audit next, the
audit page collapses into a single-column layout.

This is because in the `save` action we manually render the `show`
view, bypassing the layout that gets set in the `show` action.

This change fixes this by correctly setting the layout in this case.