Add foreign keys to Allocations
We currently have no referential integrity on our Allocations table.
This allows us to do things like save "anyone" as an allocation's user
ID, which then causes things to fall over.

This change adds in foreign keys from the Allocations table to the
Content Items and Users tables.

In order to add this foreign key, we also need to add a unique index
to the users' `uid` column (which is also a good thing!).