Restrict iframe proxy endpoint to SAMEORIGIN frames
We currently have a constraint on our iframe proxy endpoint that means
a user must be logged in to access it. However, when we deploy our app
to Heroku, we don't have Signon, and this constraint fails.

We can't just enable the endpoint on Heroku, because this would allow
arbitrary people to use our proxy in their own iframes.

This change allows us to make this endpoint work on Heroku, whilst also
preventing people from using our proxy on their own sites. We achieve
this by:

- removing the logged-in constraint
- making our own proxy return an `X-FRAME-OPTIONS SAMEORIGIN` header,
  which is exactly the thing we were trying to circumvent on GOV.UK in
  the first place, and therefore should be adequate security when
  applied here, too