Disable JavaScript for multiple select tests
A couple of our JavaScript-enabled tests appear to fail intermittently
on Jenkins. We have tried to fix these in the past, but it seems the fix
didn't work.

This change instead disables JavaScript in these tests, which should
remove all race conditions and make the tests more reliable. The
JavaScript for these elements is already tested elsewhere in this suite,
and in the JavaScript unit tests.