Pad factory-generated content titles
Our Content Item factory currently generates titles like:

```
content-item-title-1
content-item-title-2
content-item-title-3
...
content-item-title-10
```

This can cause problems in tests which assert that the page does not
contain `content-item-title-1` when the page contains
`content-item-title-10` (of which `content-item-title-1` is a substring).

This change fixes this by zero-padding the numbers to generate titles
like:

```
content-item-title-0001
content-item-title-0002
content-item-title-0003
...
content-item-title-0010
```

This is in keeping with the other indices in this factory, and
guarantees that none of the IDs are substrings of each other.