Add Dimensions::Item model
The grain for our first facts table is the `content item` as all
requests/queries from the business are related to retrieving information
regarding the performance of content items filtered/grouped by different
criteria.

For example:

- Number of content items published per organisation
- Number of content items in PDF format for an organisation
- Number of visits in the last six months for content items tagged to a taxon

I have added `organisation_id` as an attribute and not as a dimension
because we don’t have yet any need related to filtering by organisation.
This is going to change in the future, but for the time being it is
better to keep the code aligned with our immediate needs and don't think
in what we are going to need in the future.

I have decided to validate only the `content_id` because, as of today,
there are content_items with no content_id returned by `rummager`; this
is going to change in incoming PRs when we move to other sources of data.